refactor: change users to user for consistency
