+++ backend/.env.sample
@@ -1,5 +1,7 @@
-MONGODB_URI=
-DB_NAME=
+MONGODB_URI=your_mongo_uri
+DB_NAME=your_db_name
 CORS_ORIGIN=*
 PORT=8080
 NODE_ENV=development
+JWT_ACCESS_SECRET=your_access_secret
+JWT_ACCESS_EXPIRY=1d
