+++ backend/src/app.js
@@ -3,7 +3,7 @@ import express from 'express';
 
 import passport from './middleware/auth.js';
 import globalErrorHandler from './middleware/globalErrorHandler.js';
-import { UserRoutes } from './routes/users.route.js';
+import { UserRoutes } from './routes/user.route.js';
 
 const app = express();
 
@@ -20,7 +20,7 @@ app.use(express.urlencoded({ extended: true }));
 app.use(passport.initialize());
 
 app.use('/api/auth', UserRoutes);
-app.use('/api/users', UserRoutes);
+app.use('/api/user', UserRoutes);
 
 app.use(globalErrorHandler);
