+++ backend/src/middleware/globalErrorHandler.js
@@ -47,7 +47,7 @@ const globalErrorHandler = (error, req, res, next) => {
       '\x1b[31m',
       '\n\n-------------------- ERROR --------------------\n',
       message,
-      '\n -----------------------------------------------\n'
+      '\n-----------------------------------------------\n'
     );
   }
