+++ backend/src/repositories/UserRepository.js
@@ -1,5 +1,5 @@
 import BaseRepository from './BaseRepository.js';
-import { User } from '../models/users.model.js';
+import { User } from '../models/user.model.js';
 import ApiError from '../utils/ApiError.js';
 import HTTP_STATUS from '../utils/httpStatus.js';
